requirements.txt
# Pin NumPy < 2.0 because many binary wheels (incl. FAISS) are not yet compiled against NumPy 2.x
numpy<2.0
faiss-cpu==1.8.0
numpy>=1.23
sentence-transformers==2.7.0
openai>=1.40.0
python-dotenv>=1.0.1
tqdm>=4.66.4
streamlit==1.50.0